+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/grid/WirelessGridContainerMenu.java
@@ -1,7 +1,8 @@
 package com.refinedmods.refinedstorage2.platform.common.containermenu.grid;
 
+import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.api.grid.Grid;
-import com.refinedmods.refinedstorage2.platform.common.containermenu.slot.PlayerSlotReference;
+import com.refinedmods.refinedstorage2.platform.api.item.SlotReference;
 import com.refinedmods.refinedstorage2.platform.common.content.Menus;
 
 import net.minecraft.network.FriendlyByteBuf;
@@ -10,16 +11,16 @@ import net.minecraft.world.entity.player.Inventory;
 public class WirelessGridContainerMenu extends AbstractGridContainerMenu {
     public WirelessGridContainerMenu(final int syncId, final Inventory playerInventory, final FriendlyByteBuf buf) {
         super(Menus.INSTANCE.getWirelessGrid(), syncId, playerInventory, buf);
-        this.disabledPlayerInventorySlot = PlayerSlotReference.of(buf);
+        this.disabledSlot = PlatformApi.INSTANCE.getSlotReference(buf).orElse(null);
         addSlots(0);
     }
 
     public WirelessGridContainerMenu(final int syncId,
                                      final Inventory playerInventory,
                                      final Grid grid,
-                                     final PlayerSlotReference slotReference) {
+                                     final SlotReference slotReference) {
         super(Menus.INSTANCE.getWirelessGrid(), syncId, playerInventory, grid);
-        this.disabledPlayerInventorySlot = slotReference;
+        this.disabledSlot = slotReference;
         addSlots(0);
     }
 }