+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/ClientModInitializerImpl.java
@@ -38,6 +38,7 @@ import java.util.List;
 import com.mojang.blaze3d.platform.InputConstants;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
@@ -287,6 +288,13 @@ public class ClientModInitializerImpl extends AbstractClientModInitializer imple
             GLFW.GLFW_KEY_TAB,
             KEY_BINDINGS_TRANSLATION_KEY
         )));
+        KeyMappings.INSTANCE.setOpenWirelessGrid(KeyBindingHelper.registerKeyBinding(new KeyMapping(
+            createTranslationKey("key", "open_wireless_grid"),
+            InputConstants.Type.KEYSYM,
+            InputConstants.UNKNOWN.getValue(),
+            KEY_BINDINGS_TRANSLATION_KEY
+        )));
+        ClientTickEvents.END_CLIENT_TICK.register(client -> handleInputEvents());
     }
 
     private void registerModelPredicates() {