+++ CHANGELOG.md
@@ -7,9 +7,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+-   "Open Wireless Grid" keybinding.
+-   Curios integration on Forge.
+-   Trinkets integration on Fabric.
+
 ### Changed
 
--  You can now recharge the Controller in item form.
+-   You can now recharge the Controller in item form.
 
 ### Removed